new_math_site/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
boto3==1.17.102
botocore==1.20.102
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.4
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-nested-admin==3.3.3
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
dynaconf==3.1.4
gunicorn==20.1.0
hvac==0.10.14
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
loguru==0.5.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.0
python-monkey-business==1.0.0
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
shortuuid==1.0.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
validators==0.18.2
whitenoise==5.3.0
win32-setctime==1.0.3
